+++ test/util_test.dart
@@ -2,7 +2,7 @@ library github.test.util_test;
 
 import "helper.dart";
 
-import "package:github/common.dart";
+import "package:github/src/common.dart";
 import "package:test/test.dart";
 
 main() {